# backend/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Watertracker",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Watertracker operations"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://XXXXXXX.onrender.com"
    }
  ],
...
  "paths": {
...
    "/water": {
      "post": {
        "tags": [
          "water"
        ],
        "summary": "Add water",
        "operationId": "addWater",
        "description": "Add water to user intake with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json:": {
              "schema": {
                "type": "object",
                "required": [
                  "volume",
                  "date",
                  "userId"
                ],
                "properties": {
                  "date": {
                    "description": "Date of water intake record",
                    "type": "string",
                    "example": "2024-05-08T16:12:14"
                  },
                  "volume": {
                    "description": "Volume of water intake record",
                    "type": "number",
                    "example": 500
                  },
                  "userId": {
                    "description": "Id of a user who this water record is belongs to",
                    "type": "string",
                    "example": "65e4decdd286b30065d54af9"
                  },
                  "_id": {
                    "description": "Id of water log",
                    "type": "string",
                    "example": "66b750163beb29c60d215f8c"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added water",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully added water!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          }
        }
      }
    },
    "/water/{id}": {
      "patch": {
        "tags": [
          "water"
        ],
        "summary": "Update water",
        "operationId": "updateWater",
        "description": "Update water with payload operation description",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json:": {
              "schema": {
                "type": "object",
                "required": [
                  "volume",
                  "date",
                  "userId"
                ],
                "properties": {
                  "date": {
                    "description": "Date of water intake record",
                    "type": "string",
                    "example": "2024-05-08T16:12:14"
                  },
                  "volume": {
                    "description": "Volume of water intake record",
                    "type": "number",
                    "example": 500
                  },
                  "userId": {
                    "description": "Id of a user who this water record is belongs to",
                    "type": "string",
                    "example": "65e4decdd286b30065d54af9"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated water",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated water data!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "tags": [
          "water"
        ],
        "summary": "Delete water by id",
        "operationId": "deleteWater",
        "description": "Delete water by id operation description",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted water log"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "water": {
        "type": "object",
        "required": [
          "date",
          "volume",
          "userId"
        ],
        "properties": {
          "date": {
            "description": "Date of water intake record",
            "type": "string",
            "example": "2024-05-08T16:12:14.954208"
          },
          "volume": {
            "description": "Volume of water intake record",
            "type": "number",
            "example": 500
          },
          "userId": {
            "description": "Id of a user who this water record is belongs to",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "_id": {
            "description": "Id of water log",
            "type": "string",
            "example": "66b750163beb29c60d215f8c"
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Bad request"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Not found"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
    }
  }
}